Avoid crashing when external-dependent service fails
As it was, the application would fail on startup if the external service used to retrieve area information was down (or changed something in their interface). With this change, the exception that is the cause will be logged, but the application will still start as usual. It does mean no assessment areas will be returned anymore, but other functionality of the search application should still work.