AER-1309 Fix startup problem (#23)
* Avoid crashing when external-dependent service fails
As it was, the application would fail on startup if the external service used to retrieve area information was down (or changed something in their interface). With this change, the exception that is the cause will be logged, but the application will still start as usual. It does mean no assessment areas will be returned anymore, but other functionality of the search application should still work.
* Fixed the Natura2000WfsInterpreter
By the looks of it the definition for http://geodata.nationaalgeoregister.nl/inspire/ps-natura2000/ has changed, and it's now returning MultGeometry instead of MultiSurface. Not sure when/how/why, but this change seems to fix the exception at least.