AER-1587 adjust pdok urls (#39)
* Updated Natura2000WfsInterpreter to work with new URL
The new URL uses a different schema (because why not), so adjusted accordingly. This means that the old URL won't work anymore (supporting both felt unnecessary, since the new one is already live).
On the plus side, it's simpler.
For some reason they also changed SRID from ETRS89 (srid 4258)  to ETRS89-extended (srid 3035).
* Removed old test files
These were from the old URL, so might as well remove them now.
* Updated pdok locatieserver URL
Also made it configurable with a default.
* Ensure correct code is called
If you have a protected method for testing purposes, you better use it in the public methods.