Attempt at improving suggestions from Bing map search
The locations endpoint has a query part, but that doesn't really work as suggestions: it only returns results that are pretty much exactly what you're searching for, and something like 'edin' won't result anything.
There is a autosuggest endpoint, but that one doesn't return geo information... So tying the 2 together, first the suggestion for sensible suggstions and then the locations endpoint to get geo information. Does mean more roundtrips, but there's a max of 10 results anyway.