Adjusted second call to include name information
The name does not really fit in the locations-with-adress endpoint (when using it along with countryregion, you'll get just United Kingdom as a result...), so have to use the locations-with-query endpoint (which is the same, but has different query parameters, not the cleanest endpoint definition ever).
Does mean more results are returned, so reduced number of max results.