AER-2539 Attempt at improving suggestions from Bing map search (#43)
The locations endpoint has a query part, but that doesn't really work as suggestions: it only returns results that are pretty much exactly what you're searching for, and something like 'edin' won't result anything.
There is a autosuggest endpoint, but that one doesn't return geo information... So tying the 2 together, first the suggestion for sensible suggestions and then the locations endpoint to get geo information. Does mean more roundtrips, but there's a max number of results anyway.
The name does not really fit in the locations-with-adress endpoint (when using it along with countryregion, you'll get just United Kingdom as a result...), so have to use the locations-with-query endpoint (which is the same, but has different query parameters, not the cleanest endpoint definition ever) when that is present, as it's the most qualifying property by the looks of it.