Add retry functionality in BingSearchService
There's a max requests per second Bing supports with non-enterprise keys (and probably also with enterprise keys, but it should be higher in that case).
If we hit that, wait a second before trying again. 1 second is pretty long, but as we're using the async process anyway, I think it's OK to do this(?)